feature(exports): extracted `createGroqBuilder` methods to separate files
